build: crdb-api-client workflow
Define a new Github workflow to build and publish
`crdb-api-client` NPM package. It requires two steps to ensure
that package version is incremented before publishing to the
registry and when produced outputs contain changes comparing
to previous version.
First workflow run compares if there's changes comparing to
previously published version and if yes it checks that `version`
is incremented in `package.json` file.
If version is the same as already published, then it creates a
new PR with incremented version in `package.json` and developer
should approve/merge PR to proceed. After PR is merged,
this workflow invoked once again, it checks that package has
changes **and** package version is incremented so now it is
possible to publish package to NPM registry.

Release note: None